fix(prr): forensic 100/100 remediation across P0-P4 with verified migrations and tests
- P0/P1: Add Razorpay webhook to financial idempotency and fix outer-join row-locking
- P0/P1: Reconcile adoption partial unique index and kennel capacity with FOR UPDATE locks
- P1: Enforce facility IDOR checks across shelter service endpoints
- P1: Enforce quarantine requirements for adoptability, adoption kennels, and transfers
- P1: Remove medical mock names and enforce real DB queries with 404 validation
- P1: Implement post-rescue feedback survey job in ARQ worker cron
- P1: Remove Alembic SQL interceptor hack and fix f61eaab29ac9 invalid column indexes
- P1: Add production Prometheus alerting rules for latency, 5xx, DB/Redis, worker, kennel capacity
- P1: Scrub leaked credentials from .env.example and enforce MFA for admin roles by default
- P2: Upgrade donation money fields to Decimal and mask donor PII (PAN, address, name)
- P2: Add rescue_case_id and shelter_facility_id foreign keys to finance expenses
- P2: Add BCS 1-9 check constraint to clinical exams (migration b7c8d9e0f1a2)
- P2: Implement sibling adoption application rejection on foster-to-adopt conversion
- P2: Optimize agent availability queries with batch geopos and Redis pipeline
- P2: Add reviewer audit trail to lost & found match resolutions
- Pass all ruff check, ruff format, mypy, unit (1335 tests), and regression suites

+++ src/pawguard/modules/donation/models.py
@@ -3,6 +3,7 @@
 import uuid
 from datetime import date as date_type
 from datetime import datetime
+from decimal import Decimal
 from enum import StrEnum
 from typing import TYPE_CHECKING
 
@@ -102,7 +103,7 @@ class DogSponsorship(UUIDPkMixin, TimestampMixin, AuditMixin, Base):
         nullable=False,
         index=True,
     )
-    monthly_amount: Mapped[float] = mapped_column(Numeric(10, 2), nullable=False)
+    monthly_amount: Mapped[Decimal] = mapped_column(Numeric(10, 2), nullable=False)
     currency: Mapped[str] = mapped_column(String(3), default="USD", nullable=False)
     status: Mapped[SponsorshipStatus] = mapped_column(
         String(32), default=SponsorshipStatus.ACTIVE, nullable=False, index=True
@@ -131,7 +132,7 @@ class RecurringSubscription(UUIDPkMixin, TimestampMixin, AuditMixin, Base):
         nullable=False,
         index=True,
     )
-    amount: Mapped[float] = mapped_column(Numeric(10, 2), nullable=False)
+    amount: Mapped[Decimal] = mapped_column(Numeric(10, 2), nullable=False)
     currency: Mapped[str] = mapped_column(String(3), default="USD", nullable=False)
     frequency: Mapped[RecurringFrequency] = mapped_column(
         String(32), default=RecurringFrequency.MONTHLY, nullable=False
@@ -185,7 +186,7 @@ class Donation(UUIDPkMixin, TimestampMixin, AuditMixin, Base):
         index=True,
     )
 
-    amount: Mapped[float] = mapped_column(Numeric(10, 2), nullable=False)
+    amount: Mapped[Decimal] = mapped_column(Numeric(10, 2), nullable=False)
     currency: Mapped[str] = mapped_column(String(3), default="USD", nullable=False)
     donation_type: Mapped[DonationType] = mapped_column(
         String(32), default=DonationType.ONE_TIME, nullable=False
@@ -225,7 +226,7 @@ class DonationCampaign(UUIDPkMixin, TimestampMixin, SoftDeleteMixin, AuditMixin,
 
     name: Mapped[str] = mapped_column(String(128), nullable=False)
     description: Mapped[str | None] = mapped_column(Text, nullable=True)
-    target_amount: Mapped[float] = mapped_column(Numeric(12, 2), nullable=False)
+    target_amount: Mapped[Decimal] = mapped_column(Numeric(12, 2), nullable=False)
     currency: Mapped[str] = mapped_column(String(3), default="USD", nullable=False)
     campaign_type: Mapped[CampaignType] = mapped_column(
         String(32), default=CampaignType.GENERAL, nullable=False
